// BB-main/src/bb/generated/BbInput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BallBearingGeometry } from "./BallBearingGeometry";
import type { BallBearingKind } from "./BallBearingKind";
import type { BbOperatingConditions } from "./BbOperatingConditions";
import type { BbSolverParams } from "./BbSolverParams";
import type { Material } from "./Material";

/**
 * 최상위 입력 래퍼.
 */
export type BbInput = { 
/**
 * 볼베어링 변종. 기본값 `Acbb` (§3.6.1.3)
 */
kind: BallBearingKind, geometry: BallBearingGeometry, material: Material, operating: BbOperatingConditions, solver: BbSolverParams, };